requirements.txt
streamlit
pandas
plotly
scikit-learn
matplotlib
seaborn
Pillow
